EMC VMAX database lock for symconfigure
When VMAX is executing a symconfigure command it database stay
in lock to another execute and need to wait for this.
The STGAdm will wait 60 seconds and try again. It will happen
more one time (3 tentatives on total)
For this the return code when is diffrent of 0 was also improved

Change-Id: Id42939153bf8536e03f782f038b2abea655fbf14