LUN TAG is not correct using two pools.
When the total number of LUNs is 'even' two pools are necessary.
On the first pool the LUN name is created correct using TAG_LUN_LSSID,
but on the second pool the name was been created using LUN_LSSID.
Example with TAG 'QA':
QA_LUN_0A45 on first pool
LUN_0B45 on second pool

This problem is fixed on this review.
Example with TAG 'QA':
QA_LUN_0A45 on first pool
QA_LUN_0B45 on second pool

Change-Id: Ibf21c058159d15b5ea481244c513bb290bdab399